Add info on specific nearby MTA bus lines
This includes listing out the line in particular, as well as direct
links to information pages (schedules, route maps) on the MTA site.
There are way more buses that are nearby, but I limited my selection to
those buses landing within one block of the venue.

The text around the MTA bus line was likewise updated.