Remove note about subway+light rail convenience
I hadn't honestly looked at the directions directly, but just assumed
that the light rail was much closer than the subway. But really, it just
shaves a couple minutes off of walking, not the whole distance. It's
certainly not enough to specifically call it out as a really good
option.