Add details about the Light Rail
This also winds up adding a smidgen of detail about the airport, but I
don't think any more needs to be said about it than that. There are
really only two *good* ways to get from BWI into the part of Baltimore
where we're having the conference (driving and light rail). While I'm
not being hyper-explicit on that point, as it's really more of an
opinion than a hard fact, I think this serves as implication enough.