Rework driving directions/intro section
There was a fair amount of information duplication, probably due to
copying over from a different original source. That has now been
removed. Additionally, there was some extra data about other locations
at the University, which for likely similar reasons was kept, but is not
necessary for our purposes. It has also been removed.

Finally, the link information for the navigation was moved up to be near
to the address, and both Google and Apple maps have been included. Links
like this were added without mention before, but in brief, I include
them as some people (though I am not one of them) do prefer them,
perhaps because of the tighter integration with Siri and other iOS/Mac
OS X features.