Add information about the CCC
This is the most recent addition to Baltimore's public transportation
offerings, so it likely didn't yet make it into the materials that we
received. But, it's a nice (and free!) way to get from place to place in
the city, and it's right next door to our venue, so it's got a mention
now.