Properly URL-encode link addresses
This is a mark in the column of "can't believe I didn't know that", but
ampersands in URLs in "href" attributes should, like all ampersands
anywhere in HTML documents, be escaped as "&amp;". I first thought this
might be the case when seeing the previously-existing and
computer-generated Google Maps embed code, and a bit of research** seems
to have confirmed that.

I don't believe any "ambiguity" resulted from the links, and they worked
fine, but as I feel we should all strive to meet all applicable
standards, this has now been fixed.

** http://stackoverflow.com/a/3705601