Add newspost for Infinity Interactive
I also added some styling so that I could include the quote they gave us
as a block quote. It winds up looking like an email forward, which may
or may not be a good thing. If anyone with better design sense than me
wants to take a whack at this, that probably wouldn't be the worst idea.