Add OmniTI donation!
This news item looks a bit different than the others due to the higher
level of sponsorship we received, per our prospectus. The primary
difference is that the font is bigger on the title, and the image has a
larger footprint (roughly 9k px vs 3k). That's about as big as I feel
like I could make the difference without sending it over the edge from
"more prominent" to "too much"/"comical".