Add Standards of Conduct page/link
The SOC was already referenced by the registration process, but it's
important to have these things front-and-center, if only because people
aren't saving the registration form for reference later at our event.

This is a very initial/basic design of this. The text comes from
previous years' workshops/conferences, and the provenance traces back to
the Geek Feminism's template, and the layout attempts to mimic the home
page's design as much as possible.

I opted for the gray background, because as cool as the parallax effect
is and how much I love it, it's a lot of specific (in an almost legal
sense) text to read and that may be difficult with layered text behind
it.

There are a couple spots of inline CSS added to the index page to get
buttons side-by-side, and on the SOC to remove extra height to support
the index page's parallax effect. Inline CSS is definitely a code smell
to me, but without fully understanding the implications of making
changes in the CSS files proper, and wanting to get this change in as
quickly as possible, I went ahead and did it. #YOLO